requirements.txt
PyQt5
mss
pillow
numpy
Flask
Flask-SocketIO
eventlet